E2E Github Action workflow for minikube

+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/internal/readiness/Readiness.java
@@ -114,7 +114,7 @@ public class Readiness {
     }
 
     return spec.getReplicas().intValue() == status.getReplicas() &&
-      spec.getReplicas().intValue() <= status.getAvailableReplicas();
+      spec.getReplicas() <= status.getAvailableReplicas();
   }
